pyproject.toml
[project]
name = "strava-fit"
version = "0.1.0"
description = ""
authors = [
    {name = "Phoon",email = ""}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "fit-tool (>=0.9.13,<0.10.0)",
    "pytz (>=2025.1,<2026.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
